libs/shared/src/DataPipelineService/pyproject.toml
[tool.poetry]
name = "datapipelineservice"
version = "0.1.0"
description = "Data Pipeline for Synapse"
authors = ["snyata <core@synavate.tech>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
motor = "^3.3.2"
python-dotenv = "^1.0.1"
pymongo = "^4.6.1"
llama-index = "^0.9.41"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
